Updated DrawingRegister Controller Deleted InvitationVM Updated SeedData

+++ DrawingRegisterWeb/Controllers/DrawingRegistersController.cs
@@ -8,6 +8,16 @@ using Microsoft.EntityFrameworkCore;
 
 namespace DrawingRegisterWeb
 {
+	/* DrawingRegisters - Creates a unique identifier;
+	 
+	 * DrawingRegisterUsers - Creates a relationship with DrawingRegisters and AspNetUsers;
+	 
+	 * Invitations - Enable AspNetUsers to join DrawingRegisterUsers by mutual agreement between two or more users;
+	 
+	 * TODO: RefreshSignIn - likely a poor approach. There should be a better way to update roles while the user is connected.
+			 Creates a bug where you cannot use two or more accounts from the same device at the same time.
+			 RefreshSignIn changes the login for all open tabs of the most recently used account.*/
+
 	[Authorize]
 	public class DrawingRegistersController : Controller
 	{
@@ -25,15 +35,23 @@ namespace DrawingRegisterWeb
 			_signInManager = signInManager;
 		}
 
+
+
+
 		public async Task<IActionResult> Index()
 		{
+			//RegisterVM gathers current user's DrawingRegisterUsers (if user has DrawingRegister) and Invitations
 			var registerVM = new RegisterVM();
 			var user = await _userManager.GetUserAsync(User);
 			var userRegister = await _context.DrawingRegisterUsers.FirstOrDefaultAsync(u => u.UserId == user.Id);
 
 			if (userRegister == null)
 			{
-				registerVM.Invitations = await _context.Invitations.Include(s => s.Status).Where(i => i.UserId == user.Id).ToListAsync();
+				registerVM.Invitations = await _context.Invitations
+					.Include(s => s.Status)
+					.Include(u => u.IdentityUser)
+					.Where(i => i.UserId == user.Id || i.RecipientEmail == user.Email).ToListAsync();
+
 				return View(registerVM);
 			}
 
@@ -52,6 +70,7 @@ namespace DrawingRegisterWeb
 
 
 
+		//Create new DrawingRegister and RegisterUser for current user
 		public IActionResult Create()
 		{
 			return View();
@@ -59,7 +78,7 @@ namespace DrawingRegisterWeb
 
 		[HttpPost]
 		[ValidateAntiForgeryToken]
-		public async Task<IActionResult> Create(DrawingRegister drawingRegister)
+		public async Task<IActionResult> Create(DrawingRegister drawingRegister, bool seedData = false)
 		{
 			var user = await _userManager.GetUserAsync(User);
 
@@ -75,17 +94,34 @@ namespace DrawingRegisterWeb
 					Role = ConstData.Role_Admin_Name
 				};
 
-				//Seed required data for new drawing register - default states
+				//Seed required data for new Drawingregister - default states
 				var states = SeedDataRuntime.CreateProjectStates(drawingRegister.Id);
 
 				await _context.ProjectState.AddRangeAsync(states);
 				await _context.SaveChangesAsync();
 
-				//Seed example data for new drawing register: projects, drawings, layouts
-				var projects = SeedDataRuntime.CreateProjects(states[0], states[1], drawingRegister.Id);
+				//Seed example data for new Drawingregister: projects, drawings, layouts, documentation
+				if (seedData)
+				{
+					var projects = SeedDataRuntime.CreateProjects(states[0], states[1]);
 
-				await _context.Project.AddRangeAsync(projects);
-				await _context.SaveChangesAsync();
+					await _context.Project.AddRangeAsync(projects);
+					await _context.SaveChangesAsync();
+
+					var drawings = SeedDataRuntime.CreateDrawings(projects[0]);
+
+					await _context.Drawing.AddRangeAsync(drawings);
+					await _context.SaveChangesAsync();
+
+					var drawingFiles = SeedDataRuntime.CreateDrawingFiles(drawings[0], drawings[1], drawings[2]);
+					var layouts = SeedDataRuntime.CreateLayouts(projects[0]);
+					var documentation = SeedDataRuntime.CreateDocumentation(projects[0]);
+
+					await _context.DrawingFile.AddRangeAsync(drawingFiles);
+					await _context.Layout.AddRangeAsync(layouts);
+					await _context.Documentation.AddRangeAsync(documentation);
+					await _context.SaveChangesAsync();
+				}
 
 				await _context.AddAsync(drawingRegisterUser);
 				await _userManager.AddToRoleAsync(user!, ConstData.Role_Admin_Name);
@@ -101,6 +137,7 @@ namespace DrawingRegisterWeb
 
 
 
+		//Delete DrawingRegister and all DrawingRegisterUsers of current user's DrawingRegister
 		public async Task<IActionResult> Delete(int? id)
 		{
 			if (id == null || _context.DrawingRegisters == null)
@@ -136,7 +173,7 @@ namespace DrawingRegisterWeb
 				_context.DrawingRegisters.Remove(drawingRegister);
 			}
 
-			//Remove all roles for all users in this current drawing register
+			//Remove all roles for all users in this current Drawingregister
 			var drawingRegisterUsers = from d in _context.DrawingRegisterUsers where d.DrawingRegisterId == id select d;
 			var registerUsersList = await drawingRegisterUsers.ToListAsync();
 			var allUsers = from u in _userManager.Users select u;
@@ -156,6 +193,7 @@ namespace DrawingRegisterWeb
 
 
 
+		//Change role of choosen user in current DrawingRegister
 		[HttpPost]
 		[ValidateAntiForgeryToken]
 		[Authorize(Roles = ConstData.Role_Admin_Name)]
@@ -164,6 +202,17 @@ namespace DrawingRegisterWeb
 			var user = await _userManager.GetUserAsync(User);
 			var drawingRegisterUser = await _context.DrawingRegisterUsers.FindAsync(id);
 			var thisUser = await _userManager.FindByIdAsync(drawingRegisterUser!.UserId);
+			var adminRegisterUsers = await _context.DrawingRegisterUsers
+				.Where(d => d.DrawingRegisterId == drawingRegisterUser.DrawingRegisterId &&
+							d.Role == ConstData.Role_Admin_Name).ToListAsync();
+
+			if (drawingRegisterUser.Role == ConstData.Role_Admin_Name && adminRegisterUsers.Count == 1)
+			{
+				TempData["AdminChangeRole"] = "If you are the only one administrator, you can't change your role. " +
+					"Please assign someone as administrator.";
+
+				return RedirectToAction(nameof(Index));
+			}
 
 			if (drawingRegisterUser != null && role != null && thisUser != null)
 			{
@@ -187,6 +236,7 @@ namespace DrawingRegisterWeb
 
 
 
+		//Remove choosen user from DrawingRegister
 		[HttpPost]
 		[ValidateAntiForgeryToken]
 		[Authorize(Roles = ConstData.Role_Admin_Name)]
@@ -208,8 +258,10 @@ namespace DrawingRegisterWeb
 
 
 
+		//Leave DrawingRegister for current user
 		[HttpPost]
 		[ValidateAntiForgeryToken]
+		[Authorize(Roles = $"{ConstData.Role_Admin_Name},{ConstData.Role_Mech_Name},{ConstData.Role_Engr_Name}")]
 		public async Task<IActionResult> Leave()
 		{
 			var user = await _userManager.GetUserAsync(User);
@@ -217,6 +269,18 @@ namespace DrawingRegisterWeb
 
 			if (drawingRegisterUser != null)
 			{
+				var adminRegisterUsers = await _context.DrawingRegisterUsers
+				.Where(d => d.DrawingRegisterId == drawingRegisterUser.DrawingRegisterId &&
+							d.Role == ConstData.Role_Admin_Name).ToListAsync();
+
+				if (drawingRegisterUser.Role == ConstData.Role_Admin_Name && adminRegisterUsers.Count == 1)
+				{
+					TempData["AdminLeave"] = "If you are the only one administrator, you can't leave the register. " +
+						"Please assign someone as administrator.";
+
+					return RedirectToAction(nameof(Index));
+				}
+
 				await _userManager.RemoveFromRoleAsync(user!, drawingRegisterUser.Role);
 				_context.DrawingRegisterUsers.Remove(drawingRegisterUser);
 				await _context.SaveChangesAsync();
@@ -230,6 +294,7 @@ namespace DrawingRegisterWeb
 
 
 
+		//Create request to join existing DrawingRegister for user that has no DrawingRegister
 		public IActionResult RequestInvitation()
 		{
 			return View();
@@ -242,7 +307,7 @@ namespace DrawingRegisterWeb
 			var user = await _userManager.GetUserAsync(User);
 			var registerUser = await _userManager.Users.
 				FirstOrDefaultAsync(u => u.NormalizedEmail == invitation.RecipientEmail.ToUpper());
-			var status = await _context.Statuses.FirstOrDefaultAsync(s => s.Name == ConstData.Status_Approval_pending);
+			var status = await _context.Statuses.FirstOrDefaultAsync(s => s.Name == ConstData.Status_Request);
 
 			if (registerUser != null)
 			{
@@ -274,17 +339,13 @@ namespace DrawingRegisterWeb
 				return RedirectToAction(nameof(Index));
 			}
 
-			var invitationVM = new InvitationVM()
-			{
-				Invitation = invitation
-			};
-
-			return View(invitationVM);
+			return View(invitation);
 		}
 
 
 
 
+		//Remove request to join existing DrawingRegister for user that has no DrawingRegister
 		[HttpPost]
 		[ValidateAntiForgeryToken]
 		public async Task<IActionResult> RemoveRequestInvitation(int id)
@@ -304,6 +365,7 @@ namespace DrawingRegisterWeb
 
 
 
+		//Remove request or invitation which was assign to current user's DrawingRegister
 		[HttpPost]
 		[ValidateAntiForgeryToken]
 		[Authorize(Roles = ConstData.Role_Admin_Name)]
@@ -324,6 +386,7 @@ namespace DrawingRegisterWeb
 
 
 
+		//Accept request or invitation which was assign to current user's DrawingRegister
 		[HttpPost]
 		[ValidateAntiForgeryToken]
 		[Authorize(Roles = ConstData.Role_Admin_Name)]
@@ -349,6 +412,90 @@ namespace DrawingRegisterWeb
 			await _context.SaveChangesAsync();
 
 			return RedirectToAction(nameof(Index));
+		}
+
+
+
+
+		//Accept request or invitation which was assign to current user, who has no DrawingRegister
+		[HttpPost]
+		[ValidateAntiForgeryToken]
+		public async Task<IActionResult> AcceptInvitation(int id)
+		{
+			var user = await _userManager.GetUserAsync(User);
+			var invitation = await _context.Invitations.FindAsync(id);
+
+			if (invitation != null && user != null)
+			{
+				var drawingRegisterUser = new DrawingRegisterUsers
+				{
+					DrawingRegisterId = invitation.DrawingRegisterId,
+					UserId = user.Id,
+					Role = invitation.Role
+				};
+
+				_context.Invitations.Remove(invitation);
+				_context.DrawingRegisterUsers.Add(drawingRegisterUser);
+				await _userManager.AddToRoleAsync(user, invitation.Role);
+			}
+
+			await _context.SaveChangesAsync();
+			await _signInManager.RefreshSignInAsync(user!);
+
+			return RedirectToAction(nameof(Index));
+		}
+
+
+
+
+		//Create invitation to join current user's DrawingRegister for user that has no DrawingRegister
+		[Authorize(Roles = ConstData.Role_Admin_Name)]
+		public IActionResult Invitation()
+		{
+			return View();
+		}
+		[HttpPost]
+		[ValidateAntiForgeryToken]
+		[Authorize(Roles = ConstData.Role_Admin_Name)]
+		public async Task<IActionResult> Invitation(Invitation invitation)
+		{
+			var user = await _userManager.GetUserAsync(User);
+			var registerUser = await _userManager.Users.
+				FirstOrDefaultAsync(u => u.NormalizedEmail == invitation.RecipientEmail.ToUpper());
+			var status = await _context.Statuses.FirstOrDefaultAsync(s => s.Name == ConstData.Status_Invitation);
+
+			if (registerUser != null)
+			{
+				var hasDrawingRegisterUsers = await _context.DrawingRegisterUsers.
+					FirstOrDefaultAsync(d => d.UserId == registerUser.Id);
+
+				if (hasDrawingRegisterUsers != null)
+				{
+					ModelState.AddModelError("UserRegister",
+						"This user already has drawing register.");
+				}
+				else
+				{
+					var thisDrawingRegisterUsers = await _context.DrawingRegisterUsers.FirstOrDefaultAsync(d => d.UserId == user.Id);
+					invitation.StatusId = status!.Id;
+					invitation.DrawingRegisterId = thisDrawingRegisterUsers!.DrawingRegisterId;
+				}
+			}
+			else
+			{
+				ModelState.AddModelError("NoUser",
+					"There is no such user. Please verify that the user's e-mail is correct.");
+			}
+
+			if (ModelState.IsValid)
+			{
+				_context.Add(invitation);
+				_context.SaveChanges();
+
+				return RedirectToAction(nameof(Index));
+			}
+
+			return View(invitation);
 		}
 	}
 }
